feat: Enable users to signup or login with an email address and password
Note that Parse will make sure that the email string is a valid address.  To
check that the new features work as intended, a new user was created.  Error
messages are displayed if either text field is blank or if the user types an
incorrect password.  If Parse finds that the email is invalid, a descriptive
error message will also be displayed.