feat: Enable a user to view the feeds of his or her followers
All posts made by all followed users will be displayed in a custom
table cell and presented to the "following" user.

Note that Parse will not allow an image larger than 10mb to be posted.  The
current workaround is to compress the image (in jpeg format) but the file size
itself is not explicitly checked.